requirements.txt
flask==3.1.0
pandas==2.2.3
openpyxl==3.1.5
google-genai>=1.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
psycopg2-binary>=2.9.9
rapidfuzz>=3.0.0
unidecode>=1.3.0
